carers: collapse care-team and household into one /carers module with Add carer CTA
Three previously-separate surfaces all called "care team" made
adding a carer non-obvious:
  - /care-team (clinical visit log)
  - /household (Anchor account members, just-built)
  - /settings#care-team (local contact registry)

This collapses them into a single /carers module with one big
"Add carer" CTA at the top, two sections below, and a footer link
to the visit log:

  - On Anchor — household members + pending invites
  - Local contacts — phone numbers and notes for clinicians /
    services that don't need Anchor accounts
  - Footer link → /care-team (visit log, demoted from primary nav)

What moved:
- Renamed /household → /carers
- Extracted local-contact registry into LocalContactsSection on
  /carers; stripped its duplicate Anchor-invite UI (that lives only
  in InviteCarerFlow now)
- Replaced /settings CareTeamSection with a "Manage carers →" stub
  pointing to /carers
- Nav: replaced the /care-team primary entry with /carers labeled
  "Carers" (UserPlus icon); /care-team stays at its URL but is
  off the primary nav
- All dashboard cards (InviteFamilyCard, PendingInvitesCard,
  RecentlyAcceptedCard) and the /family HouseholdHeader now link
  to /carers
- Added nav.carers i18n keys; repointed nav.care_team to
  "Visit log" / "就诊记录"

Mental model: Carers = PEOPLE involved in care. Add carer = invite
them. Visit log = INTERACTIONS, distinct concept, separate URL.

Verified: 692 tests pass; typecheck, lint, build green.
