+++ src/components/dashboard/invite-family-card.tsx
@@ -46,7 +46,7 @@ export function InviteFamilyCard() {
           </div>
         </div>
         <Link
-          href="/household"
+          href="/carers"
           className="inline-flex items-center gap-0.5 text-[12px] text-ink-500 hover:text-ink-900"
         >
           {L("Set up", "开始")}
